Gauss noise and poisson noise together
