+++ lib/generate.py
@@ -45,11 +45,8 @@ def poissonData(N = 700, f=9, phase=0, offset=500, modulation=50):
 
 def gaussData(N=700, f=9, phase=0, offset=500, modulation=50,  noise_std=0.35*50):
     olddata, param = generateData(N=N, f = f, phase=phase)
-    data = []
-    for i in range(len(olddata)):
-        data.append(olddata*modulation+offset)
+    data = (olddata*modulation+offset)
     noise = np.random.normal(0, noise_std, size=len(data))
     noisy_data = data + noise
     intdata = np.rint(noisy_data).astype(int)
-    
-    return intdata, param
+    return intdata, param, olddata
